+++ CHANGELOG.md
@@ -17,8 +17,13 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - ability in `json-prune` scriptlet to match `key` with specific `value`
   and remove `array`/`object` if it contains specific `item` [#183]
 
+### Fixed
+
+- `prevent-eval-if` and `prevent-bab` scriptlets, now `eval.toString()` call returns original value [#481]
+
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.1.4...HEAD
 [#183]: https://github.com/AdguardTeam/Scriptlets/issues/183
+[#481]: https://github.com/AdguardTeam/Scriptlets/issues/481
 
 ## [v2.1.4] - 2025-01-20